chore(docs): fixed tests and lint
